+++ src/loader/chart/types.ts
@@ -1,0 +1,8 @@
+import { Dayjs } from 'dayjs';
+
+export type Interval = {
+  baseId: number;
+  quoteId: number;
+  date: Dayjs;
+  value: number;
+};
